fixed: created a variable instead of using a string for future use & create subfolders for settings and home subpages
